backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
flask-mysqldb
mysqlclient
mysql-connector-python
